@@ -4,25 +4,25 @@ publish_to: 'none'
 version: 0.1.0
 
 environment:
-  sdk: '>=2.19.6 <3.7.12'
+  sdk: ^3.7.2
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-      
+
   http:
   shared_preferences:
   flutter_svg:
   slidable_button:
   table_calendar:
   dropdown_below:
-  bloc: ^7.2.1
-  flutter_bloc: ^7.3.1
-  dio: ^4.0.4
+  bloc: ^9.0.0
+  flutter_bloc: ^9.1.1
+  dio: ^5.8.0+1
   flutter_slidable:
-  pin_code_fields: ^7.3.0
+  pin_code_fields: ^8.0.1
   flutter_custom_clippers: ^2.1.0
   sqflite:
   flutter_zoom_drawer: ^3.0.0
@@ -32,11 +32,12 @@ dependencies:
   flutter_switch:
   google_sign_in:
   flutter_facebook_auth:
+  hugeicons:
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^5.0.0
 
 flutter:
   uses-material-design: true
