@@ -1,4 +1,4 @@
-package com.example.alamal
+package com.alan.alamal
 
 import io.flutter.embedding.android.FlutterActivity
 
