@@ -3,44 +3,48 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-class DemoLocalization {
+class AppLocalization {
   final Locale locale;
-  DemoLocalization(this.locale);
-  static DemoLocalization? of(context) {
-    return Localizations.of<DemoLocalization>(context, DemoLocalization);
+
+  AppLocalization(this.locale);
+
+  static AppLocalization? of(context) {
+    return Localizations.of<AppLocalization>(context, AppLocalization);
   }
 
-   Map<String, String>? _localizationValues;
+  Map<String, String>? _localizationValues;
 
   Future<bool> load() async {
-    String jsonStringValues =
-        await rootBundle.loadString('lib/language/${locale.languageCode}.json');
+    String jsonStringValues = await rootBundle.loadString(
+      'lib/language/${locale.languageCode}.json',
+    );
 
     Map<String, dynamic> mappedJson = json.decode(jsonStringValues);
-    _localizationValues =
-        mappedJson.map((key, value) => MapEntry(key, value.toString()));
+    _localizationValues = mappedJson.map(
+      (key, value) => MapEntry(key, value.toString()),
+    );
     return true;
   }
 
-   String? getTranslatedValue(String key) {
+  String? getTranslatedValue(String key) {
     return _localizationValues![key];
   }
 
-  static const LocalizationsDelegate<DemoLocalization> delegate =
+  static const LocalizationsDelegate<AppLocalization> delegate =
       _DemoLocalizationDelegate();
 }
 
-class _DemoLocalizationDelegate
-    extends LocalizationsDelegate<DemoLocalization> {
+class _DemoLocalizationDelegate extends LocalizationsDelegate<AppLocalization> {
   const _DemoLocalizationDelegate();
+
   @override
   bool isSupported(Locale locale) {
     return ['ar', 'en'].contains(locale.languageCode);
   }
 
   @override
-  Future<DemoLocalization> load(Locale locale) async {
-    DemoLocalization localization = DemoLocalization(locale);
+  Future<AppLocalization> load(Locale locale) async {
+    AppLocalization localization = AppLocalization(locale);
     await localization.load();
     return localization;
   }
