@@ -9,9 +9,7 @@ class CacheHelper {
     sharedPreferences = await SharedPreferences.getInstance();
   }
 
-  static dynamic getData({
-    required String key,
-  }) {
+  static dynamic getData({required String key}) {
     return sharedPreferences.get(key);
   }
 
@@ -31,9 +29,7 @@ class CacheHelper {
     return await sharedPreferences.setDouble(key, value);
   }
 
-  static Future<bool> removeData({
-    required String key,
-  }) async {
+  static Future<bool> removeData({required String key}) async {
     return await sharedPreferences.remove(key);
   }
 
