@@ -4,75 +4,24 @@ class ColorsData {
   final String title;
   final int color;
 
-  ColorsData({
-    required this.title,
-    required this.color,
-  });
+  ColorsData({required this.title, required this.color});
 }
 
 List<ColorsData> colorsData = [
-  ColorsData(
-    title: 'Teal',
-    color: 0xFF009688,
-  ),
-  ColorsData(
-    title: 'Deep Teal',
-    color: 0xFF008080,
-  ),
-  ColorsData(
-    title: 'Green',
-    color: 0xFF4CAF50,
-  ),
-  ColorsData(
-    title: 'Emerald Green',
-    color: 0xff2E8B57,
-  ),
-  ColorsData(
-    title: 'Healing Mint',
-    color: 0xFF84C585,
-  ),
-  ColorsData(
-    title: 'Deep Purple',
-    color: 0xFF673AB7,
-  ),
-  ColorsData(
-    title: 'Calm Lavender',
-    color: 0xFFA48BD3,
-  ),
-  ColorsData(
-    title: 'Trustworthy Navy',
-    color: 0xff283593,
-  ),
-  ColorsData(
-    title: 'Navy Blue',
-    color: 0xFF000080,
-  ),
-  ColorsData(
-    title: 'Slate Blue',
-    color: 0xFF6A5ACD,
-  ),
-  ColorsData(
-    title: 'Deep Blue',
-    color: 0xff003366,
-  ),
-  ColorsData(
-    title: 'Steel Blue',
-    color: 0xff4682B4,
-  ),
-  ColorsData(
-    title: 'Hospital Blue',
-    color: 0xff5D9CEC,
-  ),
-  ColorsData(
-    title: 'Serene Aqua',
-    color: 0xff4DD0E1,
-  ),
-  ColorsData(
-    title: 'Gentle Gray',
-    color: 0xFFA3BAC5,
-  ),
-  ColorsData(
-    title: 'Medical Burgundy',
-    color: 0xFF800020,
-  ),
+  ColorsData(title: 'Teal', color: 0xFF009688),
+  ColorsData(title: 'Deep Teal', color: 0xFF008080),
+  ColorsData(title: 'Green', color: 0xFF4CAF50),
+  ColorsData(title: 'Emerald Green', color: 0xff2E8B57),
+  ColorsData(title: 'Healing Mint', color: 0xFF84C585),
+  ColorsData(title: 'Deep Purple', color: 0xFF673AB7),
+  ColorsData(title: 'Calm Lavender', color: 0xFFA48BD3),
+  ColorsData(title: 'Trustworthy Navy', color: 0xff283593),
+  ColorsData(title: 'Navy Blue', color: 0xFF000080),
+  ColorsData(title: 'Slate Blue', color: 0xFF6A5ACD),
+  ColorsData(title: 'Deep Blue', color: 0xff003366),
+  ColorsData(title: 'Steel Blue', color: 0xff4682B4),
+  ColorsData(title: 'Hospital Blue', color: 0xff5D9CEC),
+  ColorsData(title: 'Serene Aqua', color: 0xff4DD0E1),
+  ColorsData(title: 'Gentle Gray', color: 0xFFA3BAC5),
+  ColorsData(title: 'Medical Burgundy', color: 0xFF800020),
 ];
