@@ -20,12 +20,14 @@ import 'package:flutter_svg/flutter_svg.dart';
 class RegisterWithGoogle extends StatefulWidget {
   final String name;
   final String email;
+  final String photoUrl;
 
   const RegisterWithGoogle({
-    Key? key,
+    super.key,
     required this.name,
     required this.email,
-  }) : super(key: key);
+    required this.photoUrl,
+  });
 
   @override
   _RegisterWithGoogleState createState() => _RegisterWithGoogleState();
@@ -39,15 +41,11 @@ class _RegisterWithGoogleState extends State<RegisterWithGoogle> {
 
   final List<DropdownMenuItem> _dropdownTestItem = [];
   var _selectedTest;
+
   List<DropdownMenuItem> buildDropdownTestItems(List testList) {
     List<DropdownMenuItem> items = [];
     for (var i in testList) {
-      items.add(
-        DropdownMenuItem(
-          value: i,
-          child: Text(i['gender']),
-        ),
-      );
+      items.add(DropdownMenuItem(value: i, child: Text(i['gender'])));
     }
     return items;
   }
@@ -79,8 +77,8 @@ class _RegisterWithGoogleState extends State<RegisterWithGoogle> {
     });
     WidgetsBinding.instance.addPostFrameCallback((_) {
       testList.addAll([
-        {'gender': DemoLocalization.of(context)?.getTranslatedValue('male')},
-        {'gender': DemoLocalization.of(context)?.getTranslatedValue('female')},
+        {'gender': AppLocalization.of(context)?.getTranslatedValue('male')},
+        {'gender': AppLocalization.of(context)?.getTranslatedValue('female')},
       ]);
       _dropdownTestItem.addAll(buildDropdownTestItems(testList));
     });
@@ -89,380 +87,465 @@ class _RegisterWithGoogleState extends State<RegisterWithGoogle> {
 
   var formKey = GlobalKey<FormState>();
   var scaffoldKey = GlobalKey<ScaffoldState>();
+
   @override
   Widget build(BuildContext context) {
     return BlocProvider(
       create: (context) => RegisterCubit(),
       child: BlocConsumer<RegisterCubit, RegisterStates>(
         listener: (context, state) {},
-        builder: (context, state) =>
-            BlocBuilder<ThemeBloc, ThemeState>(builder: (context, state) {
-          return SafeArea(
-            child: Scaffold(
-              key: scaffoldKey,
-              body: Center(
-                child: SingleChildScrollView(
-                  child: Stack(
-                    alignment: Alignment.center,
-                    children: [
-                      SizedBox(
-                        width: double.infinity,
-                        child: SvgPicture.asset(
-                          "assets/imgs/Group6.svg",
-                          color: state.color,
-                          fit: BoxFit.fill,
-                        ),
-                      ),
-                      Form(
-                        key: formKey,
-                        child: Padding(
-                          padding: const EdgeInsets.symmetric(
-                            vertical: 26.0,
-                          ),
-                          child: Flex(
-                            direction: Axis.vertical,
-                            children: [
-                              Container(
-                                width: fullWidth(context) * 0.8,
-                                decoration: BoxDecoration(
-                                  borderRadius: BorderRadius.circular(30.0),
-                                  boxShadow: [
-                                    BoxShadow(
-                                      color: Colors.grey[300]!,
-                                      blurRadius: 5.0,
-                                      spreadRadius: 1.0,
-                                      offset: const Offset(0, 3),
-                                    ),
-                                  ],
-                                  color: Colors.white,
+        builder:
+            (context, state) => BlocBuilder<ThemeBloc, ThemeState>(
+              builder: (context, state) {
+                return SafeArea(
+                  child: Scaffold(
+                    key: scaffoldKey,
+                    body: Center(
+                      child: SingleChildScrollView(
+                        child: Stack(
+                          alignment: Alignment.center,
+                          children: [
+                            SizedBox(
+                              width: double.infinity,
+                              child: SvgPicture.asset(
+                                "assets/imgs/Group6.svg",
+                                color: state.color,
+                                fit: BoxFit.fill,
+                              ),
+                            ),
+                            Form(
+                              key: formKey,
+                              child: Padding(
+                                padding: const EdgeInsets.symmetric(
+                                  vertical: 26.0,
                                 ),
-                                child: Padding(
-                                  padding: const EdgeInsets.all(15.0),
-                                  child: Column(
-                                    crossAxisAlignment:
-                                        CrossAxisAlignment.center,
-                                    children: [
-                                      TextFormField(
-                                        keyboardType: TextInputType.name,
-                                        textInputAction: TextInputAction.next,
-                                        cursorColor: state.color,
-                                        decoration: InputDecoration(
-                                          border: const OutlineInputBorder(
-                                            borderRadius: BorderRadius.all(
-                                                Radius.circular(20.0)),
-                                          ),
-                                          focusedBorder: OutlineInputBorder(
-                                            borderSide: BorderSide(
-                                              color: state.color,
-                                            ),
-                                            borderRadius:
-                                                BorderRadius.circular(20.0),
-                                          ),
-                                          hintStyle: const TextStyle(
-                                            color: Colors.black,
-                                          ),
-                                          enabled: false,
-                                          hintText: widget.name,
+                                child: Flex(
+                                  direction: Axis.vertical,
+                                  children: [
+                                    Container(
+                                      width: fullWidth(context) * 0.8,
+                                      decoration: BoxDecoration(
+                                        borderRadius: BorderRadius.circular(
+                                          30.0,
                                         ),
-                                      ),
-                                      const SizedBox(
-                                        height: 5.0,
-                                      ),
-                                      TextFormField(
-                                        keyboardType:
-                                            TextInputType.emailAddress,
-                                        textInputAction: TextInputAction.next,
-                                        cursorColor: state.color,
-                                        decoration: InputDecoration(
-                                          border: const OutlineInputBorder(
-                                            borderRadius: BorderRadius.all(
-                                                Radius.circular(20.0)),
+                                        boxShadow: [
+                                          BoxShadow(
+                                            color: Colors.grey[300]!,
+                                            blurRadius: 5.0,
+                                            spreadRadius: 1.0,
+                                            offset: const Offset(0, 3),
                                           ),
-                                          focusedBorder: OutlineInputBorder(
-                                            borderSide: BorderSide(
-                                              color: state.color,
-                                            ),
-                                            borderRadius:
-                                                BorderRadius.circular(20.0),
-                                          ),
-                                          hintStyle: const TextStyle(
-                                            color: Colors.black,
-                                          ),
-                                          enabled: false,
-                                          hintText: widget.email,
-                                        ),
+                                        ],
+                                        color: Colors.white,
                                       ),
-                                      const SizedBox(
-                                        height: 5.0,
-                                      ),
-                                      TextFormField(
-                                        validator: (value) {
-                                          if (value!.isEmpty) {
-                                            return "${DemoLocalization.of(context)!.getTranslatedValue('plzFillField')}";
-                                          }
-                                          return null;
-                                        },
-                                        controller: phone,
-                                        keyboardType: TextInputType.phone,
-                                        textInputAction: TextInputAction.next,
-                                        cursorColor: state.color,
-                                        decoration: InputDecoration(
-                                          border: const OutlineInputBorder(
-                                            borderRadius: BorderRadius.all(
-                                                Radius.circular(20.0)),
-                                          ),
-                                          focusedBorder: OutlineInputBorder(
-                                            borderSide: BorderSide(
-                                              color: state.color,
-                                            ),
-                                            borderRadius:
-                                                BorderRadius.circular(20.0),
-                                          ),
-                                          hintStyle: const TextStyle(
-                                            color: Colors.grey,
-                                          ),
-                                          hintText:
-                                              "${DemoLocalization.of(context)!.getTranslatedValue('phone')}",
-                                        ),
-                                      ),
-                                      const SizedBox(
-                                        height: 5.0,
-                                      ),
-                                      TextFormField(
-                                        validator: (value) {
-                                          if (value!.isEmpty) {
-                                            return "${DemoLocalization.of(context)!.getTranslatedValue('plzFillField')}";
-                                          }
-                                          return null;
-                                        },
-                                        controller: password,
-                                        obscureText: RegisterCubit.get(context)
-                                            .isPassVisible,
-                                        keyboardType:
-                                            TextInputType.visiblePassword,
-                                        textInputAction: TextInputAction.next,
-                                        cursorColor: state.color,
-                                        decoration: InputDecoration(
-                                          suffixIcon: Padding(
-                                            padding: const EdgeInsets.symmetric(
-                                              horizontal: 10.0,
-                                            ),
-                                            child: IconButton(
-                                              color: state.color,
-                                              onPressed: () {
-                                                RegisterCubit.get(context)
-                                                    .changePassVisible();
-                                              },
-                                              icon: Icon(
-                                                RegisterCubit.get(context)
-                                                    .suffix,
+                                      child: Padding(
+                                        padding: const EdgeInsets.all(15.0),
+                                        child: Column(
+                                          crossAxisAlignment:
+                                              CrossAxisAlignment.center,
+                                          children: [
+                                            TextFormField(
+                                              keyboardType: TextInputType.name,
+                                              textInputAction:
+                                                  TextInputAction.next,
+                                              cursorColor: state.color,
+                                              decoration: InputDecoration(
+                                                border:
+                                                    const OutlineInputBorder(
+                                                      borderRadius:
+                                                          BorderRadius.all(
+                                                            Radius.circular(
+                                                              20.0,
+                                                            ),
+                                                          ),
+                                                    ),
+                                                focusedBorder:
+                                                    OutlineInputBorder(
+                                                      borderSide: BorderSide(
+                                                        color: state.color,
+                                                      ),
+                                                      borderRadius:
+                                                          BorderRadius.circular(
+                                                            20.0,
+                                                          ),
+                                                    ),
+                                                hintStyle: const TextStyle(
+                                                  color: Colors.black,
+                                                ),
+                                                enabled: false,
+                                                hintText: widget.name,
                                               ),
                                             ),
-                                          ),
-                                          border: const OutlineInputBorder(
-                                            borderRadius: BorderRadius.all(
-                                                Radius.circular(20.0)),
-                                          ),
-                                          focusedBorder: OutlineInputBorder(
-                                            borderSide: BorderSide(
-                                              color: state.color,
+                                            const SizedBox(height: 5.0),
+                                            TextFormField(
+                                              keyboardType:
+                                                  TextInputType.emailAddress,
+                                              textInputAction:
+                                                  TextInputAction.next,
+                                              cursorColor: state.color,
+                                              decoration: InputDecoration(
+                                                border:
+                                                    const OutlineInputBorder(
+                                                      borderRadius:
+                                                          BorderRadius.all(
+                                                            Radius.circular(
+                                                              20.0,
+                                                            ),
+                                                          ),
+                                                    ),
+                                                focusedBorder:
+                                                    OutlineInputBorder(
+                                                      borderSide: BorderSide(
+                                                        color: state.color,
+                                                      ),
+                                                      borderRadius:
+                                                          BorderRadius.circular(
+                                                            20.0,
+                                                          ),
+                                                    ),
+                                                hintStyle: const TextStyle(
+                                                  color: Colors.black,
+                                                ),
+                                                enabled: false,
+                                                hintText: widget.email,
+                                              ),
                                             ),
-                                            borderRadius:
-                                                BorderRadius.circular(20.0),
-                                          ),
-                                          hintStyle: const TextStyle(
-                                            color: Colors.grey,
-                                          ),
-                                          hintText:
-                                              "${DemoLocalization.of(context)!.getTranslatedValue('password')}",
-                                        ),
-                                      ),
-                                      const SizedBox(
-                                        height: 5.0,
-                                      ),
-                                      TextFormField(
-                                        validator: (value) {
-                                          if (value!.isEmpty) {
-                                            return "${DemoLocalization.of(context)!.getTranslatedValue('plzFillField')}";
-                                          }
-                                          return null;
-                                        },
-                                        buildCounter: (
-                                          context, {
-                                          required currentLength,
-                                          required isFocused,
-                                          maxLength,
-                                        }) =>
-                                            null,
-                                        controller: age,
-                                        keyboardType: TextInputType.number,
-                                        textInputAction: TextInputAction.done,
-                                        maxLength: 2,
-                                        cursorColor: state.color,
-                                        decoration: InputDecoration(
-                                          border: const OutlineInputBorder(
-                                            borderRadius: BorderRadius.all(
-                                                Radius.circular(20.0)),
-                                          ),
-                                          focusedBorder: OutlineInputBorder(
-                                            borderSide: BorderSide(
-                                              color: state.color,
+                                            const SizedBox(height: 5.0),
+                                            TextFormField(
+                                              validator: (value) {
+                                                if (value!.isEmpty) {
+                                                  return "${AppLocalization.of(context)!.getTranslatedValue('plzFillField')}";
+                                                }
+                                                return null;
+                                              },
+                                              controller: phone,
+                                              keyboardType: TextInputType.phone,
+                                              textInputAction:
+                                                  TextInputAction.next,
+                                              cursorColor: state.color,
+                                              decoration: InputDecoration(
+                                                border:
+                                                    const OutlineInputBorder(
+                                                      borderRadius:
+                                                          BorderRadius.all(
+                                                            Radius.circular(
+                                                              20.0,
+                                                            ),
+                                                          ),
+                                                    ),
+                                                focusedBorder:
+                                                    OutlineInputBorder(
+                                                      borderSide: BorderSide(
+                                                        color: state.color,
+                                                      ),
+                                                      borderRadius:
+                                                          BorderRadius.circular(
+                                                            20.0,
+                                                          ),
+                                                    ),
+                                                hintStyle: const TextStyle(
+                                                  color: Colors.grey,
+                                                ),
+                                                hintText:
+                                                    "${AppLocalization.of(context)!.getTranslatedValue('phone')}",
+                                              ),
                                             ),
-                                            borderRadius:
-                                                BorderRadius.circular(20.0),
-                                          ),
-                                          hintStyle: const TextStyle(
-                                            color: Colors.grey,
-                                          ),
-                                          hintText:
-                                              "${DemoLocalization.of(context)!.getTranslatedValue('age')}",
-                                        ),
-                                      ),
-                                      const SizedBox(
-                                        height: 5.0,
-                                      ),
-                                      DropdownBelow(
-                                        icon: Icon(
-                                          Icons.keyboard_arrow_down_rounded,
-                                          color: state.color,
-                                        ),
-                                        itemWidth: 100,
-                                        boxHeight: 60,
-                                        boxPadding: const EdgeInsets.all(10.0),
-                                        boxDecoration: BoxDecoration(
-                                          borderRadius:
-                                              BorderRadius.circular(20.0),
-                                          border: Border.all(
-                                            color: Colors.grey,
-                                          ),
-                                        ),
-                                        itemTextstyle: TextStyle(
-                                          color: Colors.black,
-                                          fontFamily: CacheHelper.getData(
-                                                      key: 'languageCode') ==
-                                                  'ar'
-                                              ? tajawalMedium
-                                              : poppinsMedium,
-                                        ),
-                                        boxTextstyle: TextStyle(
-                                          fontFamily: CacheHelper.getData(
-                                                      key: 'languageCode') ==
-                                                  'ar'
-                                              ? tajawalMedium
-                                              : poppinsMedium,
-                                          color: Colors.black,
-                                        ),
-                                        onBoxStateChanged: (change) => true,
-                                        value: _selectedTest,
-                                        onChanged: onChangeDropdownTests,
-                                        items: _dropdownTestItem,
-                                        hint: Text(
-                                          "${DemoLocalization.of(context)!.getTranslatedValue('gender')}",
-                                          style: TextStyle(
-                                            color: Colors.grey[600],
-                                            fontFamily: CacheHelper.getData(
-                                                        key: 'languageCode') ==
-                                                    'ar'
-                                                ? tajawalRegular
-                                                : poppinsRegular,
-                                          ),
-                                        ),
-                                      ),
-                                      const SizedBox(
-                                        height: 20.0,
-                                      ),
-                                      Builder(
-                                        builder: (context) => !isLoadingLogin
-                                            ? CustomButton(
-                                                primaryColor: state.color,
-                                                height: 60.0,
-                                                width: fullWidth(context) * 0.8,
-                                                bgColor: state.color,
-                                                context: context,
-                                                onPressed: () {
-                                                  if (formKey.currentState!
-                                                          .validate() &&
-                                                      (_selectedTest != null)) {
-                                                    setState(() {
-                                                      isLoadingLogin = true;
-                                                    });
-                                                    Future.delayed(
-                                                      const Duration(
-                                                          seconds: 2),
-                                                      () {
-                                                        setState(() {
-                                                          isLoadingLogin =
-                                                              false;
-                                                        });
-                                                        UsersAuth()
-                                                            .userRegisterWithGoogle(
-                                                          accountType:
-                                                              'user', // 0 Is Type User || 1 Is Type Doctor
-                                                          userData: users,
-                                                          userPassword:
-                                                              password.text,
-                                                          userPhone: phone.text,
-                                                          userAge: age.text,
-                                                          userGender: setGender(
-                                                              _selectedTest[
-                                                                      'gender']
-                                                                  .toString()),
-                                                          context: context,
-                                                          mounted: mounted,
-                                                        );
-                                                      },
-                                                    );
-                                                  }
-                                                  if (_selectedTest == null) {
-                                                    CustomSnackBar(
-                                                      context: context,
-                                                      translateKey:
-                                                          'plzChooseGen',
-                                                    );
-                                                  }
-                                                },
-                                                translateKey: 'signWithUs',
-                                              )
-                                            : Container(
-                                                height: 60.0,
-                                                width: fullWidth(context) * 0.8,
-                                                alignment: Alignment.center,
-                                                child:
-                                                    CircularProgressIndicator(
-                                                  color: state.color,
+                                            const SizedBox(height: 5.0),
+                                            TextFormField(
+                                              validator: (value) {
+                                                if (value!.isEmpty) {
+                                                  return "${AppLocalization.of(context)!.getTranslatedValue('plzFillField')}";
+                                                }
+                                                return null;
+                                              },
+                                              controller: password,
+                                              obscureText:
+                                                  RegisterCubit.get(
+                                                    context,
+                                                  ).isPassVisible,
+                                              keyboardType:
+                                                  TextInputType.visiblePassword,
+                                              textInputAction:
+                                                  TextInputAction.next,
+                                              cursorColor: state.color,
+                                              decoration: InputDecoration(
+                                                suffixIcon: Padding(
+                                                  padding:
+                                                      const EdgeInsets.symmetric(
+                                                        horizontal: 10.0,
+                                                      ),
+                                                  child: IconButton(
+                                                    color: state.color,
+                                                    onPressed: () {
+                                                      RegisterCubit.get(
+                                                        context,
+                                                      ).changePassVisible();
+                                                    },
+                                                    icon: Icon(
+                                                      RegisterCubit.get(
+                                                        context,
+                                                      ).suffix,
+                                                    ),
+                                                  ),
+                                                ),
+                                                border:
+                                                    const OutlineInputBorder(
+                                                      borderRadius:
+                                                          BorderRadius.all(
+                                                            Radius.circular(
+                                                              20.0,
+                                                            ),
+                                                          ),
+                                                    ),
+                                                focusedBorder:
+                                                    OutlineInputBorder(
+                                                      borderSide: BorderSide(
+                                                        color: state.color,
+                                                      ),
+                                                      borderRadius:
+                                                          BorderRadius.circular(
+                                                            20.0,
+                                                          ),
+                                                    ),
+                                                hintStyle: const TextStyle(
+                                                  color: Colors.grey,
+                                                ),
+                                                hintText:
+                                                    "${AppLocalization.of(context)!.getTranslatedValue('password')}",
+                                              ),
+                                            ),
+                                            const SizedBox(height: 5.0),
+                                            TextFormField(
+                                              validator: (value) {
+                                                if (value!.isEmpty) {
+                                                  return "${AppLocalization.of(context)!.getTranslatedValue('plzFillField')}";
+                                                }
+                                                return null;
+                                              },
+                                              buildCounter:
+                                                  (
+                                                    context, {
+                                                    required currentLength,
+                                                    required isFocused,
+                                                    maxLength,
+                                                  }) => null,
+                                              controller: age,
+                                              keyboardType:
+                                                  TextInputType.number,
+                                              textInputAction:
+                                                  TextInputAction.done,
+                                              maxLength: 2,
+                                              cursorColor: state.color,
+                                              decoration: InputDecoration(
+                                                border:
+                                                    const OutlineInputBorder(
+                                                      borderRadius:
+                                                          BorderRadius.all(
+                                                            Radius.circular(
+                                                              20.0,
+                                                            ),
+                                                          ),
+                                                    ),
+                                                focusedBorder:
+                                                    OutlineInputBorder(
+                                                      borderSide: BorderSide(
+                                                        color: state.color,
+                                                      ),
+                                                      borderRadius:
+                                                          BorderRadius.circular(
+                                                            20.0,
+                                                          ),
+                                                    ),
+                                                hintStyle: const TextStyle(
+                                                  color: Colors.grey,
+                                                ),
+                                                hintText:
+                                                    "${AppLocalization.of(context)!.getTranslatedValue('age')}",
+                                              ),
+                                            ),
+                                            const SizedBox(height: 5.0),
+                                            DropdownBelow(
+                                              icon: Icon(
+                                                Icons
+                                                    .keyboard_arrow_down_rounded,
+                                                color: state.color,
+                                              ),
+                                              itemWidth: 100,
+                                              boxHeight: 60,
+                                              boxPadding: const EdgeInsets.all(
+                                                10.0,
+                                              ),
+                                              boxDecoration: BoxDecoration(
+                                                borderRadius:
+                                                    BorderRadius.circular(20.0),
+                                                border: Border.all(
+                                                  color: Colors.grey,
                                                 ),
                                               ),
+                                              itemTextstyle: TextStyle(
+                                                color: Colors.black,
+                                                fontFamily:
+                                                    CacheHelper.getData(
+                                                              key:
+                                                                  'languageCode',
+                                                            ) ==
+                                                            'ar'
+                                                        ? tajawalMedium
+                                                        : poppinsMedium,
+                                              ),
+                                              boxTextstyle: TextStyle(
+                                                fontFamily:
+                                                    CacheHelper.getData(
+                                                              key:
+                                                                  'languageCode',
+                                                            ) ==
+                                                            'ar'
+                                                        ? tajawalMedium
+                                                        : poppinsMedium,
+                                                color: Colors.black,
+                                              ),
+                                              onBoxStateChanged:
+                                                  (change) => true,
+                                              value: _selectedTest,
+                                              onChanged: onChangeDropdownTests,
+                                              items: _dropdownTestItem,
+                                              hint: Text(
+                                                "${AppLocalization.of(context)!.getTranslatedValue('gender')}",
+                                                style: TextStyle(
+                                                  color: Colors.grey[600],
+                                                  fontFamily:
+                                                      CacheHelper.getData(
+                                                                key:
+                                                                    'languageCode',
+                                                              ) ==
+                                                              'ar'
+                                                          ? tajawalRegular
+                                                          : poppinsRegular,
+                                                ),
+                                              ),
+                                            ),
+                                            const SizedBox(height: 20.0),
+                                            Builder(
+                                              builder:
+                                                  (context) =>
+                                                      !isLoadingLogin
+                                                          ? CustomButton(
+                                                            primaryColor:
+                                                                state.color,
+                                                            height: 60.0,
+                                                            width:
+                                                                fullWidth(
+                                                                  context,
+                                                                ) *
+                                                                0.8,
+                                                            bgColor:
+                                                                state.color,
+                                                            context: context,
+                                                            onPressed: () {
+                                                              if (formKey
+                                                                      .currentState!
+                                                                      .validate() &&
+                                                                  (_selectedTest !=
+                                                                      null)) {
+                                                                setState(() {
+                                                                  isLoadingLogin =
+                                                                      true;
+                                                                });
+                                                                Future.delayed(
+                                                                  const Duration(
+                                                                    seconds: 2,
+                                                                  ),
+                                                                  () {
+                                                                    setState(() {
+                                                                      isLoadingLogin =
+                                                                          false;
+                                                                    });
+                                                                    UsersAuth().userRegisterWithGoogle(
+                                                                      accountType:
+                                                                          'user',
+                                                                      // 0 Is Type User || 1 Is Type Doctor
+                                                                      userData:
+                                                                          users,
+                                                                      userPassword:
+                                                                          password
+                                                                              .text,
+                                                                      userPhone:
+                                                                          phone
+                                                                              .text,
+                                                                      userAge:
+                                                                          age.text,
+                                                                      userGender: setGender(
+                                                                        _selectedTest['gender']
+                                                                            .toString(),
+                                                                      ),
+                                                                      context:
+                                                                          context,
+                                                                      mounted:
+                                                                          mounted,
+                                                                    );
+                                                                  },
+                                                                );
+                                                              }
+                                                              if (_selectedTest ==
+                                                                  null) {
+                                                                CustomSnackBar(
+                                                                  context:
+                                                                      context,
+                                                                  translateKey:
+                                                                      'plzChooseGen',
+                                                                );
+                                                              }
+                                                            },
+                                                            translateKey:
+                                                                'signWithUs',
+                                                          )
+                                                          : Container(
+                                                            height: 60.0,
+                                                            width:
+                                                                fullWidth(
+                                                                  context,
+                                                                ) *
+                                                                0.8,
+                                                            alignment:
+                                                                Alignment
+                                                                    .center,
+                                                            child:
+                                                                CircularProgressIndicator(
+                                                                  color:
+                                                                      state
+                                                                          .color,
+                                                                ),
+                                                          ),
+                                            ),
+                                          ],
+                                        ),
                                       ),
-                                    ],
-                                  ),
+                                    ),
+                                    const SizedBox(height: 40.0),
+                                    CustomButton(
+                                      primaryColor: state.color,
+                                      height: 60.0,
+                                      width: fullWidth(context) * 0.75,
+                                      context: context,
+                                      translateKey: 'change_account',
+                                      titleColor: state.color,
+                                      onPressed: () async {
+                                        await GoogleSignInApi.logout();
+                                        pop(context: context);
+                                      },
+                                    ),
+                                  ],
                                 ),
                               ),
-                              const SizedBox(
-                                height: 40.0,
-                              ),
-                              CustomButton(
-                                primaryColor: state.color,
-                                height: 60.0,
-                                width: fullWidth(context) * 0.75,
-                                context: context,
-                                translateKey: 'change_account',
-                                titleColor: state.color,
-                                onPressed: () async {
-                                  await GoogleSignInApi.logout();
-                                  pop(context: context);
-                                },
-                              ),
-                            ],
-                          ),
+                            ),
+                          ],
                         ),
-                      )
-                    ],
+                      ),
+                    ),
                   ),
-                ),
-              ),
+                );
+              },
             ),
-          );
-        }),
       ),
     );
   }
