@@ -1,7 +1,6 @@
 import 'package:alamal/controller/constant.dart';
 import 'package:alamal/controller/custom_button.dart';
-import 'package:alamal/localization/localization.dart';
-import 'package:alamal/model/cacheHelper.dart';
+import 'package:alamal/model/lang_bloc.dart';
 import 'package:alamal/model/theme_bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -27,10 +26,11 @@ class _BookingPageState extends State<BookingPage> {
         return Theme(
           data: Theme.of(context).copyWith(
             colorScheme: ColorScheme.light(
-              primary: context
-                  .read<ThemeBloc>()
-                  .state
-                  .color, // header background color
+              primary:
+                  context
+                      .read<ThemeBloc>()
+                      .state
+                      .color, // header background color
               onPrimary: Colors.white, // header text color
               onSurface:
                   context.read<ThemeBloc>().state.color, // body text color
@@ -63,10 +63,11 @@ class _BookingPageState extends State<BookingPage> {
         return Theme(
           data: Theme.of(context).copyWith(
             colorScheme: ColorScheme.light(
-              primary: context
-                  .read<ThemeBloc>()
-                  .state
-                  .color, // header background color
+              primary:
+                  context
+                      .read<ThemeBloc>()
+                      .state
+                      .color, // header background color
               onPrimary: Colors.white, // header text color
               onSurface:
                   context.read<ThemeBloc>().state.color, // body text color
@@ -93,11 +94,9 @@ class _BookingPageState extends State<BookingPage> {
 
   void _confirmBooking() {
     if (selectedDate == null || selectedTime == null) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(
-          content: Text("Select Date And Time"),
-        ),
-      );
+      ScaffoldMessenger.of(
+        context,
+      ).showSnackBar(const SnackBar(content: Text("Select Date And Time")));
       return;
     }
 
@@ -106,161 +105,141 @@ class _BookingPageState extends State<BookingPage> {
     String notes = notesController.text;
   }
 
+  String? langCode;
+
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<ThemeBloc, ThemeState>(builder: (context, state) {
-      return SafeArea(
-        child: Scaffold(
-          appBar: AppBar(
-            backgroundColor: state.color,
-            elevation: 0.0,
-            leading: Directionality(
-              textDirection: CacheHelper.getData(key: 'languageCode') == 'ar'
-                  ? TextDirection.ltr
-                  : TextDirection.rtl,
-              child: IconButton(
-                onPressed: () {
-                  Navigator.pop(context);
-                },
-                color: Colors.white,
-                iconSize: 22.0,
-                icon: CacheHelper.getData(key: 'languageCode') == 'ar'
-                    ? const Icon(
-                        Icons.arrow_forward_ios_rounded,
-                      )
-                    : const Icon(
-                        Icons.arrow_back_ios_new_rounded,
+    return BlocBuilder<LanguageBloc, LanguageState>(
+      builder: (context, state) {
+        langCode = state.locale.languageCode;
+        return BlocBuilder<ThemeBloc, ThemeState>(
+          builder: (context, state) {
+            return SafeArea(
+              child: Scaffold(
+                appBar: AppBar(
+                  backgroundColor: state.color,
+                  elevation: 0.0,
+                  leading: CustomBackButton(langCode.toString(), context),
+                  toolbarHeight: 75.0,
+                  shape: const RoundedRectangleBorder(
+                    borderRadius: BorderRadius.only(
+                      bottomLeft: Radius.circular(40.0),
+                      bottomRight: Radius.circular(40.0),
+                    ),
+                  ),
+                  titleSpacing: 0.0,
+                  leadingWidth: 65.0,
+                  title: const Text(
+                    "Booking Date",
+                    overflow: TextOverflow.visible,
+                    softWrap: true,
+                    style: TextStyle(color: Colors.white, fontSize: 20.0),
+                  ),
+                ),
+                body: Padding(
+                  padding: const EdgeInsets.all(16.0),
+                  child: Column(
+                    children: [
+                      const SizedBox(height: 10.0),
+                      CustomButton(
+                        width: fullWidth(context),
+                        context: context,
+                        primaryColor: state.color,
+                        hasChild: true,
+                        onPressed: () => _selectDate(context),
+                        child: Text(
+                          selectedDate == null
+                              ? "Select Date"
+                              : intl.DateFormat(
+                                'yyyy-MM-dd',
+                              ).format(selectedDate!),
+                          style: TextStyle(color: state.color, fontSize: 18.0),
+                        ),
                       ),
-              ),
-            ),
-            toolbarHeight: 75.0,
-            shape: const RoundedRectangleBorder(
-              borderRadius: BorderRadius.only(
-                bottomLeft: Radius.circular(40.0),
-                bottomRight: Radius.circular(40.0),
-              ),
-            ),
-            titleSpacing: 0.0,
-            leadingWidth: 65.0,
-            title: const Text(
-              "Booking Date",
-              overflow: TextOverflow.visible,
-              softWrap: true,
-              style: TextStyle(
-                color: Colors.white,
-                fontSize: 20.0,
-              ),
-            ),
-          ),
-          body: Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: Column(
-              children: [
-                const SizedBox(height: 10.0),
-                CustomButton(
-                  width: fullWidth(context),
-                  context: context,
-                  primaryColor: state.color,
-                  hasChild: true,
-                  onPressed: () => _selectDate(context),
-                  child: Text(
-                    selectedDate == null
-                        ? "Select Date"
-                        : intl.DateFormat('yyyy-MM-dd').format(selectedDate!),
-                    style: TextStyle(
-                      color: state.color,
-                      fontSize: 18.0,
-                    ),
-                  ),
-                ),
-                const SizedBox(
-                  height: 16.0,
-                ),
-                CustomButton(
-                  width: fullWidth(context),
-                  context: context,
-                  primaryColor: state.color,
-                  hasChild: true,
-                  onPressed: () => _selectTime(context),
-                  child: Text(
-                    selectedTime == null
-                        ? "Select Time"
-                        : selectedTime!.format(context),
-                    style: TextStyle(
-                      color: state.color,
-                      fontSize: 18.0,
-                    ),
-                  ),
-                ),
-                const SizedBox(
-                  height: 26.0,
-                ),
-                const Align(
-                  alignment: AlignmentDirectional.centerStart,
-                  child: Text(
-                    "Notes For Doctor :",
-                    style: TextStyle(
-                      fontSize: 18,
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                ),
-                const SizedBox(
-                  height: 10.0,
-                ),
-                TextFormField(
-                  style: TextStyle(
-                    color: state.color,
-                    // fontFamily: checkTextFont(workplace.text),
-                  ),
-                  cursorWidth: 1.0,
-                  maxLines: 6,
-                  // controller: workplace,
-                  keyboardType: TextInputType.text,
-                  textInputAction: TextInputAction.done,
-                  cursorColor: state.color,
-                  decoration: InputDecoration(
-                    border: const OutlineInputBorder(
-                      borderRadius: BorderRadius.all(Radius.circular(20.0)),
-                    ),
-                    focusedBorder: OutlineInputBorder(
-                      borderSide: BorderSide(
-                        color: state.color,
+                      const SizedBox(height: 16.0),
+                      CustomButton(
+                        width: fullWidth(context),
+                        context: context,
+                        primaryColor: state.color,
+                        hasChild: true,
+                        onPressed: () => _selectTime(context),
+                        child: Text(
+                          selectedTime == null
+                              ? "Select Time"
+                              : selectedTime!.format(context),
+                          style: TextStyle(color: state.color, fontSize: 18.0),
+                        ),
                       ),
-                      borderRadius: BorderRadius.circular(20.0),
-                    ),
-                    hintText: "Enter your notes here ...",
+                      const SizedBox(height: 26.0),
+                      const Align(
+                        alignment: AlignmentDirectional.centerStart,
+                        child: Text(
+                          "Notes For Doctor :",
+                          style: TextStyle(
+                            fontSize: 18,
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
+                      ),
+                      const SizedBox(height: 10.0),
+                      TextFormField(
+                        style: TextStyle(
+                          color: state.color,
+                          // fontFamily: checkTextFont(workplace.text),
+                        ),
+                        cursorWidth: 1.0,
+                        maxLines: 6,
+                        // controller: workplace,
+                        keyboardType: TextInputType.text,
+                        textInputAction: TextInputAction.done,
+                        cursorColor: state.color,
+                        decoration: InputDecoration(
+                          border: const OutlineInputBorder(
+                            borderRadius: BorderRadius.all(
+                              Radius.circular(20.0),
+                            ),
+                          ),
+                          focusedBorder: OutlineInputBorder(
+                            borderSide: BorderSide(color: state.color),
+                            borderRadius: BorderRadius.circular(20.0),
+                          ),
+                          hintText: "Enter your notes here ...",
+                        ),
+                      ),
+                      const Spacer(),
+                      AnimatedContainer(
+                        duration: const Duration(milliseconds: 200),
+                        width:
+                            isLongPressed
+                                ? fullWidth(context) - 100.0
+                                : fullWidth(context),
+                        height: isLongPressed ? 50.0 : 60.0,
+                        child: CustomButton(
+                          onHighlightChanged: (isHighlighted) {
+                            setState(() {
+                              isLongPressed = isHighlighted;
+                            });
+                          },
+                          height: isLongPressed ? 50.0 : 60.0,
+                          width:
+                              isLongPressed
+                                  ? fullWidth(context) - 100.0
+                                  : fullWidth(context),
+                          primaryColor: state.color,
+                          context: context,
+                          onPressed: _confirmBooking,
+                          bgColor: state.color,
+                          translateKey: 'bookNow',
+                        ),
+                      ),
+                    ],
                   ),
                 ),
-                const Spacer(),
-                AnimatedContainer(
-                  duration: const Duration(milliseconds: 200),
-                  width: isLongPressed
-                      ? fullWidth(context) - 100.0
-                      : fullWidth(context),
-                  height: isLongPressed ? 50.0 : 60.0,
-                  child: CustomButton(
-                    onHighlightChanged: (isHighlighted) {
-                      setState(() {
-                        isLongPressed = isHighlighted;
-                      });
-                    },
-                    height: isLongPressed ? 50.0 : 60.0,
-                    width: isLongPressed
-                        ? fullWidth(context) - 100.0
-                        : fullWidth(context),
-                    primaryColor: state.color,
-                    context: context,
-                    onPressed: _confirmBooking,
-                    bgColor: state.color,
-                    translateKey: 'bookNow',
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      );
-    });
+              ),
+            );
+          },
+        );
+      },
+    );
   }
 }
