@@ -24,12 +24,12 @@ class _DoctorAppointmentsState extends State<DoctorAppointments> {
 
   void checkLangIfRefreshed() {
     if (CacheHelper.getData(key: 'languageCode') == 'ar' &&
-        "${DemoLocalization.of(context)?.getTranslatedValue('voice_call_appointments')}"
+        "${AppLocalization.of(context)?.getTranslatedValue('voice_call_appointments')}"
             .contains(RegExp(r'[a-zA-Z]'))) {
       debugPrint('Checking done in arabic');
       setState(() {});
     } else if (CacheHelper.getData(key: 'languageCode') == 'en' &&
-        !"${DemoLocalization.of(context)?.getTranslatedValue('voice_call_appointments')}"
+        !"${AppLocalization.of(context)?.getTranslatedValue('voice_call_appointments')}"
             .contains(RegExp(r'[a-zA-Z]'))) {
       debugPrint('Checking done in english');
       setState(() {});
@@ -38,15 +38,12 @@ class _DoctorAppointmentsState extends State<DoctorAppointments> {
 
   @override
   void initState() {
-    Future.delayed(
-      const Duration(seconds: 1),
-      () {
-        setState(() {
-          isRefreshed = true;
-          isLoading = false;
-        });
-      },
-    );
+    Future.delayed(const Duration(seconds: 1), () {
+      setState(() {
+        isRefreshed = true;
+        isLoading = false;
+      });
+    });
     super.initState();
   }
 
@@ -58,100 +55,89 @@ class _DoctorAppointmentsState extends State<DoctorAppointments> {
       builder: (context, state) {
         return Scaffold(
           key: scaffoldKey,
-          body: isEmpty == true
-              ? Center(
-                  child: Text(
-                    "${DemoLocalization.of(context)!.getTranslatedValue('noRandiYet')}",
-                    textAlign: TextAlign.center,
-                    style: const TextStyle(
-                      fontSize: 20.0,
-                      color: Colors.grey,
+          body:
+              isEmpty == true
+                  ? Center(
+                    child: Text(
+                      "${AppLocalization.of(context)!.getTranslatedValue('noRandiYet')}",
+                      textAlign: TextAlign.center,
+                      style: const TextStyle(
+                        fontSize: 20.0,
+                        color: Colors.grey,
+                      ),
                     ),
-                  ),
-                )
-              : !isLoading
+                  )
+                  : !isLoading
                   ? RefreshIndicator(
-                      onRefresh: () async {
-                        await Future.delayed(
-                          const Duration(seconds: 2),
-                          () {
-                            setState(() {
-                              isLoading = true;
-                            });
-                          },
-                        ).then((value) {
-                          Future.delayed(
-                            const Duration(seconds: 1),
-                            () {
-                              setState(() {
-                                isLoading = false;
-                              });
-                            },
-                          );
+                    onRefresh: () async {
+                      await Future.delayed(const Duration(seconds: 2), () {
+                        setState(() {
+                          isLoading = true;
                         });
+                      }).then((value) {
+                        Future.delayed(const Duration(seconds: 1), () {
+                          setState(() {
+                            isLoading = false;
+                          });
+                        });
+                      });
+                    },
+                    color: state.color,
+                    child: ListView.separated(
+                      clipBehavior: Clip.none,
+                      padding: const EdgeInsets.all(10.0),
+                      itemCount: 10,
+                      separatorBuilder:
+                          (context, index) => const SizedBox(height: 10.0),
+                      itemBuilder: (context, index) {
+                        return ListTile(
+                          onTap: () {},
+                          tileColor: state.color.withOpacity(0.15),
+                          shape: RoundedRectangleBorder(
+                            borderRadius: BorderRadius.circular(10.0),
+                          ),
+                          minLeadingWidth: 30.0,
+                          leading: Container(
+                            alignment: Alignment.center,
+                            width: 30.0,
+                            child: Icon(Icons.call, color: state.color),
+                          ),
+                          title: Text(
+                            "${AppLocalization.of(context)?.getTranslatedValue('voice_call_appointments')}",
+                            maxLines: 1,
+                            overflow: TextOverflow.ellipsis,
+                            style: TextStyle(
+                              fontFamily:
+                                  CacheHelper.getData(key: 'languageCode') ==
+                                          'en'
+                                      ? poppinsMedium
+                                      : tajawalMedium,
+                              fontSize:
+                                  CacheHelper.getData(key: 'languageCode') ==
+                                          'en'
+                                      ? 18.0
+                                      : 20.0,
+                            ),
+                          ),
+                          isThreeLine: true,
+                          subtitle: Text(
+                            'Alan Alhasan\n2025/06/25 - 12:00 AM',
+                            style: TextStyle(
+                              fontFamily:
+                                  CacheHelper.getData(key: 'languageCode') ==
+                                          'en'
+                                      ? poppinsMedium
+                                      : tajawalMedium,
+                              fontSize: 15.0,
+                            ),
+                          ),
+                        );
                       },
-                      color: state.color,
-                      child: ListView.separated(
-                        clipBehavior: Clip.none,
-                        padding: const EdgeInsets.all(10.0),
-                        itemCount: 10,
-                        separatorBuilder: (context, index) => const SizedBox(
-                          height: 10.0,
-                        ),
-                        itemBuilder: (context, index) {
-                          return ListTile(
-                            onTap: () {},
-                            tileColor: state.color.withOpacity(0.15),
-                            shape: RoundedRectangleBorder(
-                              borderRadius: BorderRadius.circular(10.0),
-                            ),
-                            minLeadingWidth: 30.0,
-                            leading: Container(
-                              alignment: Alignment.center,
-                              width: 30.0,
-                              child: Icon(
-                                Icons.call,
-                                color: state.color,
-                              ),
-                            ),
-                            title: Text(
-                              "${DemoLocalization.of(context)?.getTranslatedValue('voice_call_appointments')}",
-                              maxLines: 1,
-                              overflow: TextOverflow.ellipsis,
-                              style: TextStyle(
-                                fontFamily:
-                                    CacheHelper.getData(key: 'languageCode') ==
-                                            'en'
-                                        ? poppinsMedium
-                                        : tajawalMedium,
-                                fontSize:
-                                    CacheHelper.getData(key: 'languageCode') ==
-                                            'en'
-                                        ? 18.0
-                                        : 20.0,
-                              ),
-                            ),
-                            isThreeLine: true,
-                            subtitle: Text(
-                              'Alan Alhasan\n2025/06/25 - 12:00 AM',
-                              style: TextStyle(
-                                fontFamily:
-                                    CacheHelper.getData(key: 'languageCode') ==
-                                            'en'
-                                        ? poppinsMedium
-                                        : tajawalMedium,
-                                fontSize: 15.0,
-                              ),
-                            ),
-                          );
-                        },
-                      ),
-                    )
-                  : Center(
-                      child: CircularProgressIndicator(
-                        color: state.color,
-                      ),
                     ),
+                  )
+                  : Center(
+                    child: CircularProgressIndicator(color: state.color),
+                  ),
         );
       },
     );
