@@ -24,85 +24,76 @@ class _DoctorHomeState extends State<DoctorHome> {
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<ThemeBloc, ThemeState>(builder: (context, state) {
-      return SafeArea(
-        child: DefaultTabController(
-          length: 2,
-          child: Scaffold(
-            key: scaffoldKey,
-            appBar: AppBar(
-              toolbarHeight: 75.0,
-              backgroundColor: state.color,
-              actions: [
-                SvgPicture.asset(
-                  alamal,
-                  color: Colors.white,
-                  width: 130.0,
+    return BlocBuilder<ThemeBloc, ThemeState>(
+      builder: (context, state) {
+        return SafeArea(
+          child: DefaultTabController(
+            length: 2,
+            child: Scaffold(
+              key: scaffoldKey,
+              appBar: AppBar(
+                toolbarHeight: 75.0,
+                backgroundColor: state.color,
+                actions: [
+                  SvgPicture.asset(alamal, color: Colors.white, width: 130.0),
+                  const SizedBox(width: 24.0),
+                ],
+                leading: IconButton(
+                  onPressed: () {
+                    context.read<DrawerCubit>().toggleDrawer();
+                  },
+                  icon: const Icon(AmalIcons.menu, color: Colors.white),
                 ),
-                const SizedBox(
-                  width: 24.0,
+                bottom: TabBar(
+                  indicatorSize: TabBarIndicatorSize.tab,
+                  indicatorColor: Colors.white,
+                  tabs: [
+                    Padding(
+                      padding: const EdgeInsets.symmetric(vertical: 8.0),
+                      child: Text(
+                        "${AppLocalization.of(context)!.getTranslatedValue('consultations')}",
+                        textAlign: TextAlign.center,
+                        style: TextStyle(
+                          fontSize:
+                              CacheHelper.getData(key: 'languageCode') == 'en'
+                                  ? 18.0
+                                  : 20.0,
+                        ),
+                      ),
+                    ),
+                    Padding(
+                      padding: const EdgeInsets.symmetric(vertical: 8.0),
+                      child: Text(
+                        "${AppLocalization.of(context)!.getTranslatedValue('appointments')}",
+                        textAlign: TextAlign.center,
+                        style: TextStyle(
+                          fontSize:
+                              CacheHelper.getData(key: 'languageCode') == 'en'
+                                  ? 18.0
+                                  : 20.0,
+                        ),
+                      ),
+                    ),
+                  ],
                 ),
-              ],
-              leading: IconButton(
-                onPressed: () {
-                  context.read<DrawerCubit>().toggleDrawer();
+              ),
+              body: NotificationListener<OverscrollIndicatorNotification>(
+                onNotification: (overscroll) {
+                  overscroll.disallowIndicator();
+                  return true;
                 },
-                icon: const Icon(
-                  AmalIcons.menu,
-                  color: Colors.white,
+                child: const TabBarView(
+                  children: [DoctorConsultations(), DoctorAppointments()],
                 ),
-              ),
-              bottom: TabBar(
-                indicatorSize: TabBarIndicatorSize.tab,
-                indicatorColor: Colors.white,
-                tabs: [
-                  Padding(
-                    padding: const EdgeInsets.symmetric(vertical: 8.0),
-                    child: Text(
-                      "${DemoLocalization.of(context)!.getTranslatedValue('consultations')}",
-                      textAlign: TextAlign.center,
-                      style: TextStyle(
-                        fontSize:
-                            CacheHelper.getData(key: 'languageCode') == 'en'
-                                ? 18.0
-                                : 20.0,
-                      ),
-                    ),
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.symmetric(vertical: 8.0),
-                    child: Text(
-                      "${DemoLocalization.of(context)!.getTranslatedValue('appointments')}",
-                      textAlign: TextAlign.center,
-                      style: TextStyle(
-                        fontSize:
-                            CacheHelper.getData(key: 'languageCode') == 'en'
-                                ? 18.0
-                                : 20.0,
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-            body: NotificationListener<OverscrollIndicatorNotification>(
-              onNotification: (overscroll) {
-                overscroll.disallowIndicator();
-                return true;
-              },
-              child: const TabBarView(
-                children: [
-                  DoctorConsultations(),
-                  DoctorAppointments(),
-                ],
               ),
             ),
           ),
-        ),
-      );
-    });
+        );
+      },
+    );
   }
 }
+
 // signingConfigs {
 // debug {
 // keyAlias 'alamal'
